+++ src/test/java/main/GeneralControllerTest.java
@@ -1,10 +1,13 @@
 package main;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import main.api.request.ModerationRequest;
 import main.api.request.SettingsRequest;
 import main.controller.ApiGeneralController;
 import main.model.enums.GlobalSettingCode;
 import main.model.enums.GlobalSettingValue;
+import main.model.enums.ModerationStatus;
+import main.repository.PostRepository;
 import main.repository.SettingsRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,10 +19,13 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.web.servlet.MockMvc;
 
+import static org.hamcrest.Matchers.aMapWithSize;
 import static org.hamcrest.Matchers.hasSize;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.springframework.security.test.web.servlet.response.SecurityMockMvcResultMatchers.authenticated;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
@@ -36,6 +42,8 @@ public class GeneralControllerTest {
     private ObjectMapper objectMapper;
     @Autowired
     private SettingsRepository settingsRepository;
+    @Autowired
+    private PostRepository postRepository;
 
     @Test
     @Sql(value = {"/AddTestUsers.sql", "/AddTestPosts.sql", "/AddTags.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
@@ -130,5 +138,72 @@ public class GeneralControllerTest {
                 .andExpect(jsonPath("$.email").value("efee13413fs@gundex.hru"))
                 .andExpect(jsonPath("$.copyright").value("Nikita Stoyan"))
                 .andExpect(jsonPath("$.copyrightFrom").value("2021"));
+    }
+
+    @Test
+    @Sql(value = {"/AddTestUsers.sql", "/AddTestPosts.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
+    @Sql(value = {"/Clear.sql"}, executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
+    public void calendarTest() throws Exception {
+        this.mockMvc.perform(get("/api/calendar")
+                .queryParam("year", "2020"))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
+                .andExpect(jsonPath("$.years", hasSize(2)))
+                .andExpect(jsonPath("$.years.[0]").value("2020"))
+                .andExpect(jsonPath("$.years.[1]").value("2021"))
+                .andExpect(jsonPath("$.posts", aMapWithSize(9)))
+                .andExpect(jsonPath("$.posts.2020-03-21").value("2"))
+                .andExpect(jsonPath("$.posts.2020-01-21").value("2"))
+                .andExpect(jsonPath("$.posts.2020-09-21").value("2"))
+                .andExpect(jsonPath("$.posts.2020-10-21").value("1"))
+                .andExpect(jsonPath("$.posts.2020-11-21").value("2"));
+    }
+
+    @Test
+    @WithUserDetails("pasha@mail.ru")
+    @Sql(value = {"/AddTestUsers.sql", "/AddTestPosts.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
+    @Sql(value = {"/Clear.sql"}, executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
+    public void postModeration1() throws Exception {
+        ModerationRequest request = new ModerationRequest(1, "accept");
+        this.mockMvc.perform(post("/api/moderation")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(objectMapper.writeValueAsString(request)))
+                .andDo(print())
+                .andExpect(authenticated())
+                .andExpect(status().isForbidden());
+        assertTrue(postRepository.findById(27).isPresent());
+        assertEquals(postRepository.findById(27).get().getModerationStatus(), ModerationStatus.NEW);
+    }
+
+    @Test
+    @Sql(value = {"/AddTestUsers.sql", "/AddTestPosts.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
+    @Sql(value = {"/Clear.sql"}, executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
+    public void postModeration2() throws Exception {
+        ModerationRequest request = new ModerationRequest(1, "accept");
+        this.mockMvc.perform(post("/api/moderation")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(objectMapper.writeValueAsString(request)))
+                .andDo(print())
+                .andExpect(status().isUnauthorized());
+        assertTrue(postRepository.findById(27).isPresent());
+        assertEquals(postRepository.findById(27).get().getModerationStatus(), ModerationStatus.NEW);
+    }
+
+    @Test
+    @WithUserDetails("anna@mail.ru")
+    @Sql(value = {"/AddTestUsers.sql", "/AddTestPosts.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
+    @Sql(value = {"/Clear.sql"}, executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
+    public void postModeration3() throws Exception {
+        ModerationRequest request = new ModerationRequest(27, "accept");
+        this.mockMvc.perform(post("/api/moderation")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(objectMapper.writeValueAsString(request)))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.result").value("true"));
+
+        assertTrue(postRepository.findById(27).isPresent());
+        assertEquals(postRepository.findById(27).get().getModerationStatus(), ModerationStatus.ACCEPTED);
     }
 }
